Add google login and page authorization

+++ project.py
@@ -5,6 +5,17 @@ from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem
 
+from flask import session as login_session
+import random, string
+
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
+CLIENT_ID = json.loads(open('client_secret.json', 'r').read())['web']['client_id']
 
 #Connect to Database and create database session
 engine = create_engine('sqlite:///restaurantmenu.db')
@@ -12,6 +23,109 @@ Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+
+@app.route('/login/')
+def showLogin():
+  state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in range(32))
+  login_session['state'] = state
+  return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+  if request.args.get('state') != login_session['state']:
+    response = make_response(json.dumps('Invalid State Token!'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  code = request.data
+  try:
+    # Upgrade code for credentials object
+    oauth_flow = flow_from_clientsecrets('client_secret.json', scope='')
+    oauth_flow.redirect_uri = 'postmessage'
+    credentials = oauth_flow.step2_exchange(code)
+  except FlowExchangeError:
+    response = make_response(json.dumps('Failed to upgrade the auth code.'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Make a request to check the credential
+  access_token = credentials.access_token
+  url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+  h = httplib2.Http()
+  hreq = h.request(url, 'GET')[1]
+  #print(hreq, type(hreq))
+  result = json.loads(hreq.decode('utf-8'))
+  # Check for errors in the access
+  if result.get('error') is not None:
+    print('Here')
+    response = make_response(json.dumps(result.get('error')), 500)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  gplus_id = credentials.id_token['sub']
+  # Verify that my ID matches the access
+  if result['user_id'] !=gplus_id:
+    response = make_response(json.dumps('Token User ID does not match User'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Check to see if user is already logged in
+  stored_credentials = login_session.get('credentials')
+  stored_gplus_id = login_session.get('gplus_id')
+  if stored_credentials is not None and gplus_id == stored_gplus_id:
+    response = make_response(json.dumps('Current user already logged in'), 200)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Successful log in, store the access for later
+  login_session['credential'] = credentials.access_token
+  login_session['gplus_id'] = gplus_id
+  # Get user info
+  userInfoUrl = 'https://www.googleapis.com/oauth2/v1/userinfo'
+  params = {'access_token': credentials.access_token, 'alt': 'json'}
+  answer = requests.get(userInfoUrl, params=params)
+  data = json.loads(answer.text)
+  login_session['username'] = data['name']
+  login_session['picture'] = data['picture']
+  login_session['email'] = data['email']
+
+  # Respond
+  output = ''
+  output += '<h1>Welcome, '
+  output += login_session['username']
+  output += '!</h1>'
+  output += '<img src="'
+  output += login_session['picture']
+  output += '" style="width:300px; height:300px; border-radius: 150px;-webkit-\
+    border-radius: 150px; -moz-border-radius: 150px;">'
+  flash('You are now logged in as %s.' % login_session['username'])
+  return output
+
+
+@app.route('/gdisconnect/')
+def gdisconnect():
+  access_token = login_session.get('credential')
+  # Check if there is a user logged in
+  if access_token is None:
+    response = make_response(json.dumps('User not connected'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Request the revokation
+  url = ('https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token)
+  h = httplib2.Http()
+  result = h.request(url, 'GET')[0]
+  # If successful, delete session and respond
+  if result['status'] == '200':
+    del login_session['credential']
+    del login_session['gplus_id']
+    del login_session['username']
+    del login_session['email']
+    del login_session['picture']
+    response = make_response(json.dumps('Successfully disconnected'), 200)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Otherwise, tell client of failure
+  else:
+    response = make_response(json.dumps('Failed to revoke token'), 400)
+    response.headers['Content-Type'] = 'application/json'
+    return response
 
 
 #JSON APIs to view Restaurant Information
@@ -26,6 +140,7 @@ def restaurantMenuJSON(restaurant_id):
 def menuItemJSON(restaurant_id, menu_id):
     Menu_Item = session.query(MenuItem).filter_by(id = menu_id).one()
     return jsonify(Menu_Item = Menu_Item.serialize)
+
 
 @app.route('/restaurant/JSON')
 def restaurantsJSON():
@@ -43,6 +158,8 @@ def showRestaurants():
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+  if 'username' not in login_session:
+    return redirect('/login')
   if request.method == 'POST':
       newRestaurant = Restaurant(name = request.form['name'])
       session.add(newRestaurant)
@@ -55,6 +172,8 @@ def newRestaurant():
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       if request.form['name']:
@@ -68,6 +187,8 @@ def editRestaurant(restaurant_id):
 #Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
     session.delete(restaurantToDelete)
@@ -84,12 +205,13 @@ def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
     return render_template('menu.html', items = items, restaurant = restaurant)
-     
 
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
@@ -103,40 +225,41 @@ def newMenuItem(restaurant_id):
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
-
-    editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    if request.method == 'POST':
-        if request.form['name']:
-            editedItem.name = request.form['name']
-        if request.form['description']:
-            editedItem.description = request.form['description']
-        if request.form['price']:
-            editedItem.price = request.form['price']
-        if request.form['course']:
-            editedItem.course = request.form['course']
-        session.add(editedItem)
-        session.commit() 
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
+  if 'username' not in login_session:
+    return redirect('/login')
+  editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if request.method == 'POST':
+      if request.form['name']:
+          editedItem.name = request.form['name']
+      if request.form['description']:
+          editedItem.description = request.form['description']
+      if request.form['price']:
+          editedItem.price = request.form['price']
+      if request.form['course']:
+          editedItem.course = request.form['course']
+      session.add(editedItem)
+      session.commit() 
+      flash('Menu Item Successfully Edited')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+      return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
 
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
-    if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
-
-
+  if 'username' not in login_session:
+    return redirect('/login')
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one()
+  if request.method == 'POST':
+      session.delete(itemToDelete)
+      session.commit()
+      flash('Menu Item Successfully Deleted')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+      return render_template('deleteMenuItem.html', item = itemToDelete)
 
 
 if __name__ == '__main__':
